requirements.txt
smolagents
selenium
helium
pillow
python-dotenv 